+++ policyengine-client/src/countries/country.jsx
@@ -91,11 +91,13 @@ export default class Country {
         this.setState({
             situation: situation, 
             baselineSituationImpactIsOutdated: true,
+            situationVariationImpactIsOutdated: true,
             reformSituationImpactIsOutdated: true,
         });
     }
 
     useLocalServer = false;
+    usePolicyEngineOrgServer = false;
 }
 
 export const CountryContext = createContext(null);